Removing Waagent from default addon.yml
Race condition on reboots:  Waagent needs to be available on the node at reboot time in order to have the Azure metadata config that tells the BOSH agent which volumes/partitions need mounting for ephemeral storage, swap, etc.     

The effect of this is that neither the WAagent nor the BOSH agent come up on a reboot after you've symlinked /var/lib/waagent away.   There might be a more fine grained symlink to create to avoid storage blowups for Waagent upgrades etc,, or maybe I introduce glob/wildcard support if needed.